+++ pages/my-profile/index.page.tsx
@@ -1,6 +1,6 @@
-import ApplyHistory from "./ApplyHistory";
+import ApplyHistory from "./components/ApplyHistory";
 import Layout from "@/components/Layout";
-import Profile from "./Profile";
+import Profile from "./components/Profile";
 import styled from "@emotion/styled";
 import useProfileData from "./hooks/useProfileData";
 import useApplicationHistory from "./hooks/useApplicationHistory";
